never dream and code. focus shitbag

+++ src/VIO/rpg_svo/svo/include/svo/config.h
@@ -129,7 +129,7 @@ public:
   static int& isam2IMUFactorType() { return getInstance().isam2_imu_factor_type; }
 
   /// The type of IMU factor to use
-  static int& dt() { return getInstance().isam2_dt; }
+  static double& dt() { return getInstance().isam2_dt; }
 
 private:
   Config();
@@ -169,7 +169,7 @@ private:
   bool run_inertial_estimator;
   int isam2_n_iters;
   int isam2_imu_factor_type;
-  int isam2_dt;
+  double isam2_dt;
 };
 
 } // namespace svo
